Начало разделения main на подсистемы, вынесена часть кода
Код пока что не рабочий, требуется реализация вызова методов подсистем через сокеты. Добавлены заготовки для программы и подсистемы анализа данных.
