+++ MainAnalyseData.py
@@ -1,0 +1,10 @@
+class MainAnalyseData:
+    def __init__(self):
+        pass
+
+    def main_loop(self):
+        pass
+
+
+main = MainAnalyseData()
+main.main_loop()
